Код для генерации описаний к фото разбит на модули
